refactor(utilities): move isBlock
