+++ packages/dom/src/lib/dom.ts
@@ -14,4 +14,4 @@ export type { ElementTransform, MarkTransform } from './config/types'
 // Slate to DOM utilities
 export { convertSlate } from './utilities/convert-slate'
 export { extractCssFromStyle } from './utilities/domhandler'
-export { styleMapToAttribs } from './utilities'
+export { isEmptyObject, styleMapToAttribs } from './utilities'
