+++ packages/utilities/src/index.ts
@@ -1,3 +1,3 @@
 export * from './lib/style-object';
 export * from './lib/utilities';
-export * from './lib/blocks';
+export * from '../../html/src/lib/utilities/blocks';
